+++ src/server/actions/stripe/create-or-retrieve-customer.ts
@@ -1,0 +1,42 @@
+import { stripe } from '~/lib/stripe';
+import { createClient } from '~/utils/supabase/server';
+
+export const createOrRetrieveCustomer = async ({
+  email,
+  uuid,
+}: {
+  email: string;
+  uuid: string;
+}) => {
+  const supabaseAdmin = await createClient();
+  const { data, error } = await supabaseAdmin
+    .from('customers')
+    .select('stripe_customer_id')
+    .eq('id', uuid)
+    .single();
+
+  if (error || !data?.stripe_customer_id) {
+    const customerData: { metadata: { supabaseUUID: string }; email?: string } =
+      {
+        metadata: {
+          supabaseUUID: uuid,
+        },
+      };
+
+    if (email) customerData.email = email;
+
+    const customer = await stripe.customers.create(customerData);
+    const { error: supabaseError } = await supabaseAdmin
+      .from('customers')
+      .insert([{ id: uuid, stripe_customer_id: customer.id }]);
+
+    if (supabaseError) {
+      throw supabaseError;
+    }
+
+    console.log(`New customer created and inserted for ${uuid}`);
+    return customer.id;
+  }
+
+  return data.stripe_customer_id;
+};
